chore!: build with TypeScript v5.3.2
As TypeScript v5 detects more errors, the version bump requires to update the `Multiplicity` class.
It now only accepts number for the `max` property.
Previously, it accepted a special string value, but TypeScript v5 now complains when comparing a number with a string.
So, as the `max` property is used in number comparisons, it must only accept number.

BREAKING CHANGE: it is no more possible to pass a 'n' value for the `max` property of the `Multiplicity` class.
Pass `null` instead to have the same effect.